disaster-management-tool/requirements.txt
torch==2.0.0
fastapi==0.95.1
uvicorn==0.22.0
opencv-python==4.7.0.72
pyyaml==6.0
numpy==1.24.0
